host-agent/requirements.txt
google-adk
google-genai
uvicorn
python-dotenv
httpx
requests
a2a-sdk